LF-5298 Rename selectors file to utils
